Move rendering of the menu into the template
This allows us to get rid of the encoding of the menu as a
Mojo::ByteStream to prevent escaping. This in turn makes it possible to
write a "pure" tempalte that only takes data, no HTML. And this will
enable us to write a JSON template that takes the same data.