spell.t no longer calls load_languages
It as tricky to get to load_languages and using a shared global was also
not cool. Now %languages is a lexical in the server and no longer
accessed from the test. The test does, after all, use one of our own
dictionaries, so making sure that it can be found relative to the app
home directory had to be fixed instead. We now set the MOJO_HOME
variable in the test. This fixes the invokation via "perl t/spell.t" and
doesn't affect the invokation via "perl server.pl test" and it doesn't
affect running the server via "perl server.pl daemone", "morbo
server.pl" or "hypnotoad server.pl".